refactor: Major code improvements and Flask Blueprint architecture
- Refactor app.py into modular Flask Blueprints (routes/, utils/)
  - main.py: Dashboard and index routes
  - analysis.py: Table view, lineage, DAX, unused measures
  - insights.py: Model and report insights
  - sources.py: Source explorer
  - api.py: REST API endpoints
  - processors.py: Shared utility functions

- Fix security and code quality issues:
  - Remove BOM characters from Python files
  - Replace bare except clauses with specific exception handling
  - Add proper logging throughout the application
  - Improve API input validation and error handling

- Improve configuration (config.py):
  - Fix FLASK_ENV deprecation for Flask 2.3+
  - Add secure secret key generation with secrets module
  - Improve production configuration validation

- Enhance lineage_view.py:
  - Add comprehensive type hints
  - Implement caching for get_all_measures() and get_final_measures()
  - Add proper error handling for file operations
  - Reduce redundant file reads

- Fix whitespace issues in templates and CSS files
