+++ static/modern.css
@@ -1974,4 +1974,3 @@ tbody tr:last-child td {
   visibility: visible !important;
   opacity: 1 !important;
 }
-
