+++ templates/table_view.html
@@ -1,5 +1,5 @@
-﻿{% extends "base.html" %} 
-{% block title %}Visual Fields - Power BI Explorer{% endblock %} 
+{% extends "base.html" %}
+{% block title %}Visual Fields - Power BI Explorer{% endblock %}
 
 {% block header_icon %}<i class="fas fa-table"></i>{% endblock %}
 {% block header_title %}Visual Fields Explorer{% endblock %}
@@ -1497,5 +1497,4 @@
   });
 </script>
 {% endblock %}
-
