+++ utils/__init__.py
@@ -1,0 +1,21 @@
+"""Utility modules for the Power BI Analysis Tool."""
+
+from utils.processors import (
+    get_data_processor,
+    get_lineage_view_processor,
+    get_model_processor,
+    load_model_data,
+    extract_m_queries,
+    get_report_metrics,
+    calculate_lineage_metrics,
+)
+
+__all__ = [
+    'get_data_processor',
+    'get_lineage_view_processor',
+    'get_model_processor',
+    'load_model_data',
+    'extract_m_queries',
+    'get_report_metrics',
+    'calculate_lineage_metrics',
+]
